fix: deleteTodoByIdValidator

+++ src/api/v1/validators/todo.validator.ts
@@ -13,6 +13,6 @@ const createTodoValidator = validate([
     .withMessage('Todo Desciption must be of type string.'),
 ]);
 const deleteTodoByIdValidator = validate([
-  param('todoId').toInt().isInt().withMessage('Todo Id must be type int'),
+  param('todoId').isInt().withMessage('Todo Id must be type int'),
 ]);
 export { createTodoValidator, deleteTodoByIdValidator };
